[rename🚚️] CultureSpot-ServiceApi 모듈 추가에 따른 파일의 이동 (from Domain module to ServiceApi module)(#CUL-149)
