--- CultureSpot-Domain/src/main/java/com/culturespot/culturespotdomain/core/auth/annotation/Auth.java
@@ -1,9 +1,0 @@
-package com.culturespot.culturespotdomain.core.auth.annotation;
-
-import java.lang.annotation.*;
-
-@Target(ElementType.PARAMETER) // 파라미터에서 사용 가능
-@Retention(RetentionPolicy.RUNTIME)
-@Documented
-public @interface Auth {
-}
