--- CultureSpot-Domain/src/main/java/com/culturespot/culturespotdomain/core/auth/resolver/AuthArgumentResolver.java
@@ -1,55 +1,0 @@
-package com.culturespot.culturespotdomain.core.auth.resolver;
-
-import com.culturespot.culturespotcommon.global.exception.AuthException;
-import com.culturespot.culturespotcommon.global.exception.AuthExceptionCode;
-import com.culturespot.culturespotdomain.core.auth.annotation.Auth;
-import com.culturespot.culturespotdomain.core.user.entity.User;
-import com.culturespot.culturespotdomain.core.user.repository.UserRepository;
-import org.springframework.core.MethodParameter;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.stereotype.Component;
-import org.springframework.web.bind.support.WebDataBinderFactory;
-import org.springframework.web.context.request.NativeWebRequest;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.method.support.ModelAndViewContainer;
-
-@Component
-public class AuthArgumentResolver implements HandlerMethodArgumentResolver {
-
-    private final UserRepository userRepository;
-
-    public AuthArgumentResolver(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
-
-    @Override
-    public boolean supportsParameter(MethodParameter parameter) {
-        return parameter.hasParameterAnnotation(Auth.class) &&
-                parameter.getParameterType().equals(User.class);
-    }
-
-    @Override
-    public Object resolveArgument(
-            MethodParameter parameter,
-            ModelAndViewContainer mavContainer,
-            NativeWebRequest webRequest,
-            WebDataBinderFactory binderFactory
-    ) {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-
-        if (authentication == null || !authentication.isAuthenticated()) {
-            throw new AuthException(AuthExceptionCode.UNAUTHENTICATED_USER);
-        }
-
-        // 현재 인증된 사용자가 UserDetails인지 확인
-        Object principal = authentication.getPrincipal();
-        if (principal instanceof UserDetails userDetails) {
-            return userRepository.findByEmail(userDetails.getUsername())
-                    .orElseThrow(() -> new AuthException(AuthExceptionCode.USER_NOT_FOUND));
-        }
-
-        throw new AuthException(AuthExceptionCode.INVALID_AUTHENTICATION);
-    }
-}
