--- CultureSpot-Domain/src/main/java/com/culturespot/culturespotdomain/core/auth/resolver/ResolverRegistry.java
@@ -1,20 +1,0 @@
-package com.culturespot.culturespotdomain.core.auth.resolver;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-import java.util.List;
-
-@Configuration
-@RequiredArgsConstructor
-public class ResolverRegistry implements WebMvcConfigurer {
-    private final AuthArgumentResolver resolver;
-
-    @Override
-    public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
-        resolvers.add(resolver);
-    }
-}
-
