--- CultureSpot-Domain/src/main/java/com/culturespot/culturespotdomain/core/global/security/config/CorsConfig.java
@@ -1,34 +1,0 @@
-package com.culturespot.culturespotdomain.core.global.security.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.cors.CorsConfiguration;
-import org.springframework.web.cors.CorsConfigurationSource;
-import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
-
-import java.util.List;
-
-@Configuration
-public class CorsConfig {
-
-    @Bean
-    public static CorsConfigurationSource corsConfigurationSource() {
-        CorsConfiguration configuration = new CorsConfiguration();
-
-        configuration.setAllowedOrigins(List.of(
-                "https://localhost:3000",
-                "http://culturespot.com",
-                "https://culturespot.com"
-        ));
-        configuration.setAllowedMethods(List.of("GET", "POST", "PUT", "DELETE", "OPTIONS"));
-        configuration.setAllowedHeaders(List.of("*"));
-        configuration.setAllowCredentials(true); // 쿠키 전달 허용 여부
-        configuration.setExposedHeaders(List.of("Authorization", "Set-Cookie")); // 헤더에 추가 정보 제공
-        configuration.setMaxAge(3600L); // Pre-flight 요청 캐시 시간 설정 (초 단위)
-
-        UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
-        source.registerCorsConfiguration("/**", configuration);
-
-        return source;
-    }
-}
