--- CultureSpot-Domain/src/main/java/com/culturespot/culturespotdomain/core/global/security/filter/JwtAuthenticationFilter.java
@@ -1,56 +1,0 @@
-package com.culturespot.culturespotdomain.core.global.security.filter;
-
-import com.culturespot.culturespotdomain.core.global.security.token.JwtTokenManager;
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.stereotype.Component;
-import org.springframework.web.filter.OncePerRequestFilter;
-
-import java.io.IOException;
-
-@Component
-public class JwtAuthenticationFilter extends OncePerRequestFilter {
-    private final JwtTokenManager jwtTokenManager;
-    private final UserDetailsService userDetailsService;
-
-    public JwtAuthenticationFilter(JwtTokenManager jwtTokenManager, UserDetailsService userDetailsService) {
-        this.jwtTokenManager = jwtTokenManager;
-        this.userDetailsService = userDetailsService;
-    }
-
-    @Override
-    protected void doFilterInternal(
-            HttpServletRequest request,
-            HttpServletResponse response,
-            FilterChain filterChain
-    ) throws ServletException, IOException {
-        String token = resolveToken(request);
-
-        if (token != null && jwtTokenManager.isValidAccessToken(token)) {
-            String username = jwtTokenManager.getEmailFromAccessToken(token);
-            UserDetails userDetails = userDetailsService.loadUserByUsername(username);
-
-            UsernamePasswordAuthenticationToken authentication =
-                    new UsernamePasswordAuthenticationToken(userDetails, null, userDetails.getAuthorities());
-
-            SecurityContextHolder.getContext().setAuthentication(authentication);
-        }
-
-        filterChain.doFilter(request, response);
-    }
-
-    private String resolveToken(HttpServletRequest request) {
-        String bearerToken = request.getHeader("Authorization");
-        if (bearerToken != null && bearerToken.startsWith("Bearer ")) {
-            return bearerToken.substring(7);
-        }
-        return null;
-    }
-}
-
