+++ CultureSpot-ServiceApi/src/main/java/com/culturespot/culturespotserviceapi/core/auth/annotation/Auth.java
@@ -1,0 +1,9 @@
+package com.culturespot.culturespotserviceapi.core.auth.annotation;
+
+import java.lang.annotation.*;
+
+@Target(ElementType.PARAMETER) // 파라미터에서 사용 가능
+@Retention(RetentionPolicy.RUNTIME)
+@Documented
+public @interface Auth {
+}
