+++ CultureSpot-ServiceApi/src/main/java/com/culturespot/culturespotserviceapi/core/auth/userInfo/CustomOAuth2User.java
@@ -1,0 +1,33 @@
+package com.culturespot.culturespotserviceapi.core.auth.userInfo;
+
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.oauth2.core.user.OAuth2User;
+
+import java.util.Collection;
+import java.util.Map;
+
+public class CustomOAuth2User implements OAuth2User {
+
+    private final OAuth2User oAuth2User;
+    private final Collection<? extends GrantedAuthority> authorities;
+
+    public CustomOAuth2User(OAuth2User oAuth2User, Collection<? extends GrantedAuthority> authorities) {
+        this.oAuth2User = oAuth2User;
+        this.authorities = authorities;
+    }
+
+    @Override
+    public Map<String, Object> getAttributes() {
+        return oAuth2User.getAttributes();
+    }
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        return authorities;
+    }
+
+    @Override
+    public String getName() {
+        return oAuth2User.getAttribute("email"); // email을 PK로 사용
+    }
+}
